+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.39.0
-- **[refactor]** - Rename `useSnackBar` props `open` to `isOpen`
+## v1.39.1
+- **[fix]** - Secondary mobile menu
+- **[chore]** - Upgrade storybook to 7.2
+- **[chore]** - Upgrade vite dependencies
+- **[chore]** - Upgrade mui dependencies